Fixed Blog attachment Image and Security Panel
• Security panel now includes an IP add feature so that you don’t have
to hand write each line.
• All blog images pull from the wysiwyg image uploader.
• Will be updating the wysiwyg image uploader to something more modern
soon.

+++ includes/admin/dash_security.php
@@ -2,27 +2,43 @@
 
 <form method="post" action="">
     
-    <div class="input-field col s12">
+    <div class="input-field col s4">
+       <input name="ips" id="ips" type="text" class="" value="" placeholder="1.1.1.1"  />
+        <label for="ip">Add IP Address</label>
+    </div>
+    
+    <div class="input-field col s2">
+        <a id="ipUp" class="btn" href="#!">Add Ip</a>
+    </div>
+    
+    <div class="input-field col s6">
         <label>Authorized Ip Addresses ( Example: allow from 127.0.0.1 )</label>
         <textarea id="sec" name="sec" type="text" class="materialize-textarea form-control"><?php echo file_get_contents('includes/admin/.htaccess'); ?></textarea>
         <br>
+        <input onclick="alert('Ip Added Successfully')" type="submit" name="edited" class="btn btn-primary" value="Submit" />
     </div>
-    
-    <input onclick="alert('Ip Added Successfully')" type="submit" name="edited" class="btn btn-primary" value="Submit" />
     
 </form>
 
 
 <?php
-
-if(isset($_POST['sec'])){
-    
-    unlink('includes/admin/.htaccess');
+if(!file_exists('includes/admin/.htaccess')){    
     $file = fopen('includes/admin/.htaccess','w');
-    fwrite($file, $_POST['sec']);
+    fwrite($file,
+"order deny,allow
+deny from all
+allow from localhost" );
     fclose($file);
-    header('Location: //'.$_SERVER['HTTP_HOST'].$_SERVER['REQUEST_URI'].'');
-    
+    header('Location: //'.$_SERVER['HTTP_HOST'].$_SERVER['REQUEST_URI'].'');    
+}else{    
+        if(isset($_POST['sec'])){
+
+            unlink('includes/admin/.htaccess');
+            $file = fopen('includes/admin/.htaccess','w');
+            fwrite($file, $_POST['sec']);
+            fclose($file);
+            header('Location: //'.$_SERVER['HTTP_HOST'].$_SERVER['REQUEST_URI'].'');
+        }      
 }
 
 ?>
